requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
Brotli==1.0.9
cffi==1.14.6
click==8.0.1
colorama==0.4.4
crispy-bootstrap5==0.4
cssselect2==0.4.1
cycler==0.10.0
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
Django==3.2.4
django-autocomplete-light==3.8.2
django-braces==1.14.0
django-chartjs==2.2.1
django-cleanup==5.2.0
django-crispy-forms==1.12.0
django-easy-pdf==0.1.1
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Compress==1.10.1
fonttools==4.26.2
future==0.18.2
html5lib==1.1
importlib-metadata==4.6.2
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.1
openpyxl==3.0.7
pdfkit==0.6.1
Pillow==8.3.1
plotly==5.1.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pydyf==0.0.3
pylint==2.8.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
tenacity==8.0.1
tinycss2==1.1.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
weasyprint==53.0
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
xhtml2pdf==0.2.5
zipp==3.5.0
zopfli==0.1.8
